Stor push många klasser. Skapat Tile, Entity, Direction, BombExplosion och BombExplosionSquare, jobbat mycket på CollisionChecker också.

+++ core/src/Models/Position.java
@@ -1,18 +1,18 @@
 package Models;
 
 public class Position {
-    private final float x;
-    private final float y;
+    private final int x;
+    private final int y;
 
-    public Position(float x, float y){
+    public Position(int x, int y){
         this.x = x;
         this.y = y;
 
     }
-    public float getX(){
+    public int getX(){
         return x;
     }
-    public float getY(){
+    public int getY(){
         return y;
     }
     public Boolean isCollide(Position otherPosition){
